refactor: #5 カテゴリページコンポーネントを分割
