+++ packages/app/frontend/routes/categories/index.tsx
@@ -1,40 +1,22 @@
-import { CategoryIcon } from '@frontend/components/category/category-icon'
 import type {
   CategoryColor,
   CategoryIconType,
 } from '@frontend/components/category/types'
-import {
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-} from '@frontend/components/ui/alert-dialog'
-import { Badge } from '@frontend/components/ui/badge'
 import { Button } from '@frontend/components/ui/button'
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow,
-} from '@frontend/components/ui/table'
 import {
   ToggleGroup,
   ToggleGroupItem,
 } from '@frontend/components/ui/toggle-group'
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query'
 import { createFileRoute } from '@tanstack/react-router'
-import { Pencil, Plus, Trash2 } from 'lucide-react'
+import { Plus } from 'lucide-react'
 import type React from 'react'
 import { useState } from 'react'
 import { toast } from 'sonner'
 
+import { CategoryTable } from './-components/category-table'
 import { CreateCategoryDialog } from './-components/create-category-dialog'
+import { DeleteCategoryDialog } from './-components/delete-category-dialog'
 import { EditCategoryDialog } from './-components/edit-category-dialog'
 import {
   archiveCategory,
@@ -212,120 +194,14 @@ const CategoriesPage: React.FC = () => {
         />
       )}
 
-      <AlertDialog
-        open={deletingCategory !== null}
+      <DeleteCategoryDialog
+        category={deletingCategory}
         onOpenChange={(v) => {
           if (!v) setDeletingCategory(null)
         }}
-      >
-        <AlertDialogContent>
-          <AlertDialogHeader>
-            <AlertDialogTitle>カテゴリを削除しますか？</AlertDialogTitle>
-            <AlertDialogDescription>
-              「{deletingCategory?.name}
-              」を削除します。削除後は新規取引の選択肢に表示されなくなります。
-            </AlertDialogDescription>
-          </AlertDialogHeader>
-          <AlertDialogFooter>
-            <AlertDialogCancel>キャンセル</AlertDialogCancel>
-            <AlertDialogAction
-              onClick={() => {
-                if (deletingCategory) void handleDelete(deletingCategory)
-              }}
-            >
-              削除
-            </AlertDialogAction>
-          </AlertDialogFooter>
-        </AlertDialogContent>
-      </AlertDialog>
+        onDelete={handleDelete}
+      />
     </div>
-  )
-}
-
-type CategoryTableProps = {
-  categories: Category[]
-  onEdit: (category: Category) => void
-  onDelete: (category: Category) => void
-}
-
-const CategoryTable: React.FC<CategoryTableProps> = ({
-  categories,
-  onEdit,
-  onDelete,
-}) => {
-  if (categories.length === 0) {
-    return (
-      <p className="py-8 text-center text-sm text-muted-foreground">
-        カテゴリがありません
-      </p>
-    )
-  }
-
-  return (
-    <Table>
-      <TableHeader>
-        <TableRow>
-          <TableHead>カテゴリ名</TableHead>
-          <TableHead></TableHead>
-          <TableHead className="w-[100px]"></TableHead>
-        </TableRow>
-      </TableHeader>
-      <TableBody>
-        {categories.map((category) => (
-          <TableRow
-            key={category.id}
-            className={category.status === 'archived' ? 'opacity-50' : ''}
-          >
-            <TableCell>
-              <div className="flex items-center gap-2">
-                <CategoryIcon
-                  icon={category.icon}
-                  className="size-4 shrink-0"
-                  color={category.color}
-                />
-                {category.name}
-              </div>
-            </TableCell>
-            <TableCell>
-              <div className="flex gap-1">
-                {category.type === 'saving' && (
-                  <Badge variant="secondary">積立</Badge>
-                )}
-                {category.status === 'archived' && (
-                  <Badge variant="outline">削除済み</Badge>
-                )}
-              </div>
-            </TableCell>
-            <TableCell>
-              {category.status === 'active' && (
-                <div className="flex justify-end gap-1">
-                  <Button
-                    variant="ghost"
-                    size="icon"
-                    onClick={() => {
-                      onEdit(category)
-                    }}
-                    aria-label="編集"
-                  >
-                    <Pencil />
-                  </Button>
-                  <Button
-                    variant="ghost"
-                    size="icon"
-                    onClick={() => {
-                      onDelete(category)
-                    }}
-                    aria-label="削除"
-                  >
-                    <Trash2 />
-                  </Button>
-                </div>
-              )}
-            </TableCell>
-          </TableRow>
-        ))}
-      </TableBody>
-    </Table>
   )
 }
